Move imgages to assets folder in medulas-react-components to make them to compile
