+++ packages/medulas-react-components/src/pages/PageColumn/index.tsx
@@ -8,9 +8,9 @@ import Button from '../../components/Button';
 import Form, { useForm } from '../../components/forms/Form';
 import Hairline from '../../components/Hairline';
 import Img from '../../components/Image';
-import logo from '../assets/logo.svg';
-import logoBlack from '../assets/logoBlack.svg';
-import people from '../assets/People.svg';
+import logo from '../../theme/assets/pageColumn/logo.svg';
+import logoBlack from '../../theme/assets/pageColumn/logoBlack.svg';
+import people from '../../theme/assets/pageColumn/People.svg';
 import EmptyHeader from './EmptyHeader';
 import SubtitleSection from './SubtitleSection';
 import TitleSection from './TitleSection';
