+++ packages/medulas-react-components/src/theme/assets/iov-logo2x.png
